Move fix functions and their tests
